+++ ipi/pes/mace.py
@@ -32,7 +32,6 @@ class MACE_driver(ASEDriver):
         model,
         device="cpu",
         requires_extra: bool = False,
-        gos: bool = False,  # GPU OverSubscription
         mace_kwargs=None,
         *args,
         **kwargs,
@@ -68,9 +67,6 @@ class MACE_driver(ASEDriver):
         super().__init__(template, *args, **kwargs)
         self.requires_extra = requires_extra
 
-        if gos:
-            gpu_oversubscription()
-
     def check_parameters(self):
         """Check the arguments requuired to run the driver
 
@@ -82,44 +78,3 @@ class MACE_driver(ASEDriver):
         self.ase_calculator = MACECalculator(
             model_paths=self.model, device=self.device, **self.mace_kwargs
         )
-
-
-def gpu_oversubscription():
-    """
-    Assigns GPUs to SLURM tasks by oversubscribing when tasks exceed GPUs.
-
-    Uses SLURM_LOCALID and SLURM_NTASKS_PER_NODE to map tasks onto GPUs,
-    sets CUDA_VISIBLE_DEVICES accordingly, and prints assignment info.
-    """
-
-    import os
-
-    # Get the task's local ID on this node (used for GPU assignment and logging)
-    local_id = os.environ.get("SLURM_LOCALID", "unknown")
-
-    print("\n[GPU Oversubscription Info]")
-
-    # Get total number of GPUs available on this node
-    num_gpus = len([gpu for gpu in os.popen("nvidia-smi -L").readlines()])
-    print(f"  - Number of GPUs available: {num_gpus}")
-
-    # Parse the current task's local ID (default to 0 if not set)
-    local_id = int(os.environ.get("SLURM_LOCALID", 0))
-    print(f"  - SLURM_LOCALID: {local_id}")
-
-    # Total number of tasks assigned to this node
-    num_tasks = int(os.environ.get("SLURM_NTASKS_PER_NODE", 1))
-    print(f"  - SLURM_NTASKS_PER_NODE: {num_tasks}")
-
-    # Compute how many tasks will share each GPU
-    tasks_per_gpu = max(1, num_tasks // max(num_gpus, 1))
-    print(f"  - Estimated tasks per GPU: {tasks_per_gpu}")
-
-    # Assign a GPU ID to this task based on its local ID
-    gpu_id = local_id // tasks_per_gpu
-    gpu_id = min(gpu_id, num_gpus - 1)  # Clamp to last available GPU
-    print(f"  - Assigned GPU ID: {gpu_id}")
-
-    # Restrict visibility to the assigned GPU
-    os.environ["CUDA_VISIBLE_DEVICES"] = str(gpu_id)
-    print(f"  - CUDA_VISIBLE_DEVICES set to: {os.environ['CUDA_VISIBLE_DEVICES']}")
